adjusted field generator classes + three classes per domain size
